requirements.txt
fastapi
uvicorn
gunicorn
asyncio
bardapi
re_edge_gpt
requests
unofficial-claude2-api
easy-ernie